1.1.2 - Use Modern Objective C Syntax, fix the score type (it was in int, not float)

+++ Hachidori.h
@@ -49,7 +49,7 @@
 -(NSString *)getTotalEpisodes;
 -(int)getCurrentEpisode;
 -(BOOL)getConfirmed;
--(int)getScore;
+-(float)getScore;
 -(int)getWatchStatus;
 -(NSString *)getNotes;
 -(BOOL)getSuccess;
